+++ src/PlaywrightSharp/PaperFormat.cs
@@ -14,7 +14,7 @@ namespace PlaywrightSharp
         /// </summary>
         /// <param name="width">Page width in inches.</param>
         /// <param name="height">Page height in inches.</param>
-        public PaperFormat(double width, double height)
+        public PaperFormat(decimal width, decimal height)
         {
             Width = width;
             Height = height;
@@ -23,12 +23,12 @@ namespace PlaywrightSharp
         /// <summary>
         /// Letter: 8.5 inches x 11 inches.
         /// </summary>
-        public static PaperFormat Letter => new PaperFormat(8.5, 11);
+        public static PaperFormat Letter => new PaperFormat(8.5m, 11);
 
         /// <summary>
         /// Legal: 8.5 inches by 14 inches.
         /// </summary>
-        public static PaperFormat Legal => new PaperFormat(8.5, 14);
+        public static PaperFormat Legal => new PaperFormat(8.5m, 14);
 
         /// <summary>
         /// Tabloid: 11 inches by 17 inches.
@@ -43,47 +43,47 @@ namespace PlaywrightSharp
         /// <summary>
         /// A0: 33.1 inches by 46.8 inches.
         /// </summary>
-        public static PaperFormat A0 => new PaperFormat(33.1, 46.8);
+        public static PaperFormat A0 => new PaperFormat(33.1m, 46.8m);
 
         /// <summary>
         /// A1: 23.4 inches by 33.1 inches.
         /// </summary>
-        public static PaperFormat A1 => new PaperFormat(23.4, 33.1);
+        public static PaperFormat A1 => new PaperFormat(23.4m, 33.1m);
 
         /// <summary>
         /// A2: 16.5 inches by 23.4 inches.
         /// </summary>
-        public static PaperFormat A2 => new PaperFormat(16.54, 23.4);
+        public static PaperFormat A2 => new PaperFormat(16.54m, 23.4m);
 
         /// <summary>
         /// A3: 11.7 inches by 16.5 inches.
         /// </summary>
-        public static PaperFormat A3 => new PaperFormat(11.7, 16.54);
+        public static PaperFormat A3 => new PaperFormat(11.7m, 16.54m);
 
         /// <summary>
         /// A4: 8.27 inches by 11.7 inches.
         /// </summary>
-        public static PaperFormat A4 => new PaperFormat(8.27, 11.7);
+        public static PaperFormat A4 => new PaperFormat(8.27m, 11.7m);
 
         /// <summary>
         /// A5: 5.83 inches by 8.27 inches.
         /// </summary>
-        public static PaperFormat A5 => new PaperFormat(5.83, 8.27);
+        public static PaperFormat A5 => new PaperFormat(5.83m, 8.27m);
 
         /// <summary>
         /// A6: 4.13 inches by 5.83 inches.
         /// </summary>
-        public static PaperFormat A6 => new PaperFormat(4.13, 5.83);
+        public static PaperFormat A6 => new PaperFormat(4.13m, 5.83m);
 
         /// <summary>
         /// Page width in inches.
         /// </summary>
-        public double Width { get; set; }
+        public decimal Width { get; set; }
 
         /// <summary>
         /// Page height in inches.
         /// </summary>
-        public double Height { get; set; }
+        public decimal Height { get; set; }
 
         /// <inheritdoc cref="IEquatable{T}"/>
         public static bool operator ==(PaperFormat left, PaperFormat right)
